mejoras, funciones
