+++ .vscode/settings.json
@@ -1,31 +1,32 @@
 {
   "files.associations": {
-	"memtrack.h": "c",
-	"string.h": "c",
-	"initializer_list": "c",
-	"stdlib.h": "c",
-	"unistd.h": "c",
-	"stdio.h": "c",
-	"mt.h": "c",
-	"mtlib.h": "c",
-	"cstdlib": "c",
-	"mtstructs.h": "c",
-	"list.h": "c",
-	"mem.h": "c",
-	"stack.h": "c",
-	"a": "cpp",
-	"hash.h": "c",
-	"array": "c",
-	"string": "c",
-	"string_view": "c",
-	"ranges": "c",
-	"typeinfo": "c",
-	"compare": "c",
-	"functional": "c",
-	"tuple": "c",
-	"type_traits": "c",
-	"utility": "c"
-},
+    "memtrack.h": "c",
+    "string.h": "c",
+    "initializer_list": "c",
+    "stdlib.h": "c",
+    "unistd.h": "c",
+    "stdio.h": "c",
+    "mt.h": "c",
+    "mtlib.h": "c",
+    "cstdlib": "c",
+    "mtstructs.h": "c",
+    "list.h": "c",
+    "mem.h": "c",
+    "stack.h": "c",
+    "a": "cpp",
+    "hash.h": "c",
+    "array": "c",
+    "string": "c",
+    "string_view": "c",
+    "ranges": "c",
+    "typeinfo": "c",
+    "compare": "c",
+    "functional": "c",
+    "tuple": "c",
+    "type_traits": "c",
+    "utility": "c",
+    "utils.h": "c"
+  },
   "C_Cpp_Runner.cCompilerPath": "gcc",
   "C_Cpp_Runner.cppCompilerPath": "g++",
   "C_Cpp_Runner.debuggerPath": "gdb",
