+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+         #
+#    By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/09/02 14:34:27 by danpalac          #+#    #+#              #
-#    Updated: 2024/11/28 13:37:11 by danpalac         ###   ########.fr        #
+#    Updated: 2024/11/29 14:00:58 by danpalac         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -74,6 +74,7 @@ HASH_DIR		:= hash/
 STACK_DIR		:= stack/
 MEM_DIR			:= mem/
 MTLIB_DIR			:= mtlib/
+UTILS_DIR		:= utils/
 
 LIBFT			:= $(LIBFT_DIR)$(LIBFT_LIB)
 INCLUDES		:= $(INC)/*.h
@@ -98,20 +99,23 @@ LIST_FILES := ft_mtadd_back ft_mtfind_cmp ft_mtiter ft_mtlast ft_mtsize \
 			ft_mtadd_front ft_mtpop ft_mtpop_back ft_mtpush ft_mtswap ft_mtmigrate ft_mtreverse_rotate ft_mtrotate \
 			ft_mtinsert_index ft_mtpush_back ft_mtmigrate_back ft_mtprint \
 			ft_mtget_cmp ft_mtreplace ft_mtreplace_all ft_mtsearch_mt ft_mtadd_child \
-			ft_mtcount_type ft_mtcollect_type ft_mtset_to_free ft_mtnew_original_key \
+			ft_mtcount_type ft_mtcollect_type ft_mtnew_original_key \
 			ft_mtexists_mt ft_mtexists_list
 
 HASH_FILES := ft_mthash_insert ft_mthash_remove ft_mthash ft_mthash_find ft_mtfree_hash_table ft_mthash_new_table ft_mtprint_hash_table \
 			ft_mthash_find_index ft_mthash_add_child ft_mthash_replace_data ft_mthash_replace_key \
 			ft_mthash_table_free ft_hash_utils ft_mthash_add_node ft_mthash_exists ft_mthash_replace_node \
-			ft_mthash_count_type ft_mthash_collect_type
+			ft_mthash_count_type ft_mthash_collect_type ft_mthash_insert_child ft_mthash_new_original_key
 
-MEM_FILES :=  ft_mtnew ft_mtclear ft_mtdel_data ft_mtfree ft_mtdel_list ft_mterase ft_mtremove_data ft_mtpush_data ft_mtpush_data_back \
-			ft_mtnew_chaos ft_mtreplace_data ft_mtreplace_all_data ft_mtzip ft_mtunzip \
+MEM_FILES :=  ft_mtnew ft_mtdup ft_mtclear ft_mtdel_data ft_mtfree ft_mtdel_list ft_mterase ft_mtremove_data ft_mtpush_data ft_mtpush_data_back \
+			ft_mtnew_chaos ft_mtreplace_all_data ft_mtzip ft_mtunzip \
+
+UTILS_FILES := ft_mtadd_key_prefix ft_mtadd_key_suffix ft_mtnew_key ft_mtcmp_utils ft_mtprint_utils ft_mtutils \
+			ft_mtset_to_free
 
 #ft_stkclone.c ft_stkpush.c ft_stkpop.c ft_stkswap.c ft_stkrotate.c ft_stkreverse_rotate.c ft_stkpush_back.c ft_stkmigrate.c ft_stkinsert_index.c
 
-MTLIB_FILES:= ft_chaosmatrix ft_freedom ft_strmtdup ft_splitmt ft_submtstr ft_mtutils
+MTLIB_FILES:= ft_chaosmatrix ft_freedom ft_strmtdup ft_splitmt ft_submtstr
 
 #==========FILES==============================================================#
 
@@ -120,6 +124,7 @@ SRC_FILES+=$(addprefix $(MEM_DIR), $(MEM_FILES))
 SRC_FILES+=$(addprefix $(HASH_DIR), $(HASH_FILES))
 SRC_FILES+=$(addprefix $(MTLIB_DIR), $(MTLIB_FILES))
 SRC_FILES+=$(addprefix $(STACK_DIR), $(STACK_FILES))
+SRC_FILES+=$(addprefix $(UTILS_DIR), $(UTILS_FILES))
 
 SRCS := $(addprefix $(SRC_DIR), $(addsuffix .c, $(SRC_FILES)))
 OBJS := $(addprefix $(OBJ_DIR), $(addsuffix .o, $(SRC_FILES)))
