+++ inc/mtlib.h
@@ -13,11 +13,4 @@ char	*ft_strmtdup(const char *s1);
 char	*ft_submtstr(char const *s, unsigned int start, size_t len);
 t_mt	*ft_splitmt(const char *s, char c);
 
-// utils
-int		ft_mtcmp_key(t_mt *node, const char *key);
-int		ft_strcmp(const char *s1, const char *s2);
-void	print_title(const char *title);
-int		cmp_int(void *a, void *b, size_t n);
-void	ft_replace(t_mt *current, void *new_data, t_data_type type);
-
 #endif // MTLIB_H
