+++ src/hash/ft_mthash_insert.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_mthash_insert.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 13:22:35 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/28 12:36:57 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/11/29 14:00:04 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
