+++ src/hash/ft_mthash_insert_child.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_mthash_insert_child.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/28 14:50:14 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/28 14:50:18 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/11/29 14:00:25 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mt.h"
 
-static void	ft_mthash_insert_child(t_hash_table *ht, const char *parent_key,
+void	ft_mthash_insert_child(t_hash_table *ht, const char *parent_key,
 		const char *child_key, void *data, t_data_type type)
 {
 	t_mt	*parent;
@@ -34,8 +34,7 @@ static void	ft_mthash_insert_child(t_hash_table *ht, const char *parent_key,
 			child = ft_mtdup(parent);
 			(ft_replace(parent, NULL, BRANCH), ft_mtadd_child(parent, child));
 		}
-		child = ft_mtnew(child_key, data, type);
-		ft_mtadd_child(parent, child);
+		ft_mtadd_child(parent, ft_mtnew(child_key, data, type));
 		return ;
 	}
 	ft_mtadd_child(ft_mtnew(parent_key, NULL, BRANCH), ft_mtnew(child_key, data,
