+++ src/hash/ft_mthash_replace_data.c
@@ -6,7 +6,7 @@
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 12:46:25 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/27 13:16:56 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/11/29 14:00:15 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
