+++ src/list/ft_mtsearch_mt.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_mtsearch_mt.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 20:40:39 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/28 13:21:05 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/11/29 13:32:02 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ t_mt	*ft_mtsearch_mt(t_mt *root, const char *key, int (*cmp)(t_mt *,
 	if (cmp(root, key))
 		return (root);
 	if (root->type == BRANCH)
-		return (search_in_children(root, key, cmp)); // Buscamos en los hijos
+		return (search_in_children(root, key, cmp));
 	return (NULL);                                  
 		// Si no encontramos el nodo en ningún lado
 }
