+++ src/mem/ft_mtreplace_all_data.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_mtreplace_all_data.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 18:01:42 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/28 10:12:00 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/11/29 13:47:27 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,17 +15,14 @@
 // Replaces data in all nodes of the list 'lst' with the data 'new_data'.
 void	ft_mtreplace_all_data(t_mt **lst, void *new_data)
 {
-    t_mt	*current;
+	t_mt	*current;
 
-    if (!lst || !*lst)
-        return ;
-    current = *lst;
-    while (current)
-    {
-        ft_mtdel_by_type(&current->data, current->type);
-        current->data = new_data;
-        current = current->next;
-    }
+	if (!lst || !*lst)
+		return ;
+	current = *lst;
+	while (current)
+	{
+		ft_replace(current, new_data, current->type);
+		current = current->next;
+	}
 }
-	
-
