+++ src/utils/ft_mtprint_utils.c
@@ -1,13 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_utils.c                                         :+:      :+:    :+:   */
+/*   ft_mtprint_utils.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: danpalac <danpalac@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/28 10:21:37 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/28 10:27:30 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/29 13:27:19 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/29 13:28:00 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mt.h"
+
+void	print_title(const char *title)
+{
+	printf("\n%s\n", "==========================");
+	printf("   %s\n", title);
+	printf("%s\n\n", "==========================");
+}
