+++ src/utils/ft_mtutils.c
@@ -1,24 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_mtreplace.c                                     :+:      :+:    :+:   */
+/*   ft_mtutils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/18 09:32:46 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/29 13:56:38 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/29 13:28:54 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/29 13:46:51 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mt.h"
 
-// Replaces the node 'node' with 'new_node' in the list 'list'.
-
-void	ft_mtreplace(t_mt *node, t_mt *new_node)
+void	ft_replace(t_mt *current, void *new_data, t_data_type type)
 {
-	if (!node || !new_node)
+	if (!current)
 		return ;
-    ft_mtfree(node);
-	node = ft_mtdup(new_node);
-    ft_mtfree(new_node);
+	if (current->to_free)
+		ft_mtdel_by_type(&current->data, current->type);
+	current->data = new_data;
+	if (current->type != type)
+		current->type = type;
+	current->free_data = ft_mtget_free_data(type);
 }
